refactor: share JSX import resolution
